# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/costs:
    get:
      operationId: listOrganizationCosts
      tags: [admin]
      summary: What every account has cost to run
      description: >-
        Every organisation, invited or not, with this month's, last month's and lifetime
        running cost in real currency. Internal admins only.
      security:
        - clerkSession: []
      responses:
        "200":
          description: Cost per organisation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationCosts" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
  /admin/access:
    get:
      operationId: listAccessGrants
      tags: [admin]
      summary: Everyone who has been invited, and what they are costing
      description: >-
        The invite-only allowlist with each grant's organisation, plan, credit balance,
        and month-to-date spend in real currency. Internal admins only; this is the one
        place a cost figure is shown next to a customer.
      security:
        - clerkSession: []
      responses:
        "200":
          description: Access grants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AccessGrant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createAccessGrant
      tags: [admin]
      summary: Invite an email address
      description: >-
        Adds an email to the allowlist with the plan it should land on. Nothing is created
        until that person first logs in; then the organisation, membership, plan and credit
        balance are provisioned in one step.
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccessGrantBody" }
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
        "409":
          description: That email is already invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/access/{grantId}:
    patch:
      operationId: updateAccessGrant
      tags: [admin]
      summary: Change a grant's plan, status, or note
      description: >-
        A plan change applies to the organisation immediately for limits; the new credit
        allowance applies from next month. Suspending turns the person away at the door on
        their next request. Nothing is deleted.
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/GrantId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAccessGrantBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/access/{grantId}/credits:
    post:
      operationId: grantCredits
      tags: [admin]
      summary: Add credits to a customer by hand
      description: >-
        A trial, a goodwill top-up, or a purchase taken outside the product before billing
        exists. Written to the credit ledger with the admin's name and reason.
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/GrantId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantCreditsBody" }
      responses:
        "200":
          description: Credited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The grant has no organisation yet - the person has not logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/access/{grantId}/cost:
    get:
      operationId: getAccessGrantCost
      tags: [admin]
      summary: What one customer has cost to run this month
      description: >-
        Real spend by provider and purpose, alongside the plan price and the credit ledger.
        Admin only. This figure never reaches the customer.
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/GrantId"
      responses:
        "200":
          description: Cost detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessGrantCost" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotFound" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  schemas:
...
    AccessGrant:
      type: object
      additionalProperties: false
      required: [id, email, status, planCode, planName, initialCredits, organizationId, organizationName, clerkUserId, firstLoginAt, note, createdAt, credits, spend]
      properties:
        id: { type: string }
        email: { type: string }
        status: { type: string, enum: [invited, active, suspended] }
        planCode: { type: string }
        planName: { type: string }
        initialCredits: { type: number }
        organizationId: { type: string, nullable: true }
        organizationName: { type: string, nullable: true }
        clerkUserId: { type: string, nullable: true }
        firstLoginAt: { type: string, nullable: true }
        note: { type: string, nullable: true }
        createdAt: { type: string }
        credits:
          type: object
          additionalProperties: false
          required: [balance, monthlyAllowance]
          properties:
            balance: { type: number }
            monthlyAllowance: { type: number }
        spend:
          type: object
          additionalProperties: false
          description: Real currency. Admin-only; never included in any customer-facing shape.
          required: [monthToDateUsd, monthToDateInr, planPriceInr, planPriceUsd]
          properties:
            monthToDateUsd: { type: number }
            monthToDateInr: { type: number }
            planPriceInr: { type: number }
            planPriceUsd: { type: number }
    CreateAccessGrantBody:
      type: object
      additionalProperties: false
      required: [email, planCode]
      properties:
        email: { type: string, minLength: 3, maxLength: 320 }
        planCode: { type: string }
        organizationName: { type: string, maxLength: 120 }
        organizationId: { type: string, description: Attach to an existing organisation instead of creating one. }
        initialCredits: { type: number, minimum: 0, default: 0 }
        note: { type: string, maxLength: 2000 }
    UpdateAccessGrantBody:
      type: object
      additionalProperties: false
      properties:
        planCode: { type: string }
        status: { type: string, enum: [invited, active, suspended] }
        note: { type: string, maxLength: 2000, nullable: true }
        organizationName: { type: string, maxLength: 120 }
    GrantCreditsBody:
      type: object
      additionalProperties: false
      required: [credits, reason]
      properties:
        credits: { type: number, minimum: 1, maximum: 1000000 }
        reason: { type: string, minLength: 1, maxLength: 500 }
    OrganizationCosts:
      type: object
      additionalProperties: false
      required: [month, inrPerUsd, organizations]
      properties:
        month: { type: string }
        inrPerUsd: { type: number }
        organizations:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [organizationId, organizationName, createdAt, grantEmail, planCode, monthToDateUsd, monthToDateInr, lastMonthUsd, lastMonthInr, lifetimeUsd, lifetimeInr, wastedMonthToDateUsd, calls, lastSpendAt]
            properties:
              organizationId: { type: string }
              organizationName: { type: string }
              createdAt: { type: string }
              grantEmail: { type: string, nullable: true, description: The invited email, or null for an organisation created without an invite. }
              planCode: { type: string, nullable: true }
              monthToDateUsd: { type: number }
              monthToDateInr: { type: number }
              lastMonthUsd: { type: number }
              lastMonthInr: { type: number }
              lifetimeUsd: { type: number }
              lifetimeInr: { type: number }
              wastedMonthToDateUsd: { type: number }
              calls: { type: number }
              lastSpendAt: { type: string, nullable: true }
    AccessGrantCost:
      type: object
      additionalProperties: false
      required: [grant, month, spend, ledger]
      properties:
        grant: { $ref: "#/components/schemas/AccessGrant" }
        month: { type: string }
        spend:
          type: object
          additionalProperties: false
          required: [monthToDateUsd, todayUsd, wastedUsd, breakdown, lifetimeUsd, lifetimeInr, byMonth, byProject]
          properties:
            monthToDateUsd: { type: number }
            todayUsd: { type: number }
            wastedUsd: { type: number }
            lifetimeUsd: { type: number, description: Everything this account has ever cost to run. }
            lifetimeInr: { type: number }
            byMonth:
              type: array
              description: Cost per calendar month (UTC), newest first, up to 12 months.
              items:
                type: object
                additionalProperties: false
                required: [month, calls, costUsd, costInr]
                properties:
                  month: { type: string, description: YYYY-MM }
                  calls: { type: number }
                  costUsd: { type: number }
                  costInr: { type: number }
            byProject:
              type: array
              description: This month's cost split by project. A null projectId is spend not tied to one project.
              items:
                type: object
                additionalProperties: false
                required: [projectId, projectName, calls, costUsd, costInr]
                properties:
                  projectId: { type: string, nullable: true }
                  projectName: { type: string, nullable: true }
                  calls: { type: number }
                  costUsd: { type: number }
                  costInr: { type: number }
            breakdown:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [kind, source, outcome, calls, costUsd]
                properties:
                  kind: { type: string }
                  source: { type: string }
                  outcome: { type: string }
                  calls: { type: number }
                  costUsd: { type: number }
        ledger:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id, kind, delta, balanceAfter, description, createdAt]
            properties:
              id: { type: string }
              kind: { type: string }
              delta: { type: number }
              balanceAfter: { type: number }
              description: { type: string }
              createdAt: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  parameters:
...
    GrantId:
      name: grantId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session