# lib/api-zod/src/generated/types/organizationCosts.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationCostsOrganizationsItem } from './organizationCostsOrganizationsItem';

export interface OrganizationCosts {
  month: string;
  inrPerUsd: number;
  organizations: OrganizationCostsOrganizationsItem[];
}